Public API

Contents

Public API

Contents

Index

Public interface

Lattices

Reciprocal lattices

Cells

Index

CrystallographyCore.Cell

CrystallographyCore.Lattice

CrystallographyCore.ReciprocalLattice

CrystallographyCore.basisvectors

CrystallographyCore.basisvectors

CrystallographyCore.eachatom

CrystallographyCore.reciprocal

Public interface

Lattices

CrystallographyCore.Lattice — Type

Lattice(data::AbstractMatrix)

Construct a Lattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Examples

julia> Lattice([
           1.2 4.5 7.8
           2.3 5.6 8.9
           3.4 6.7 9.1
       ])
Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

CrystallographyCore.basisvectors — Method

basisvectors(lattice::Lattice)

Get the three basis vectors from a lattice.

Reciprocal lattices

CrystallographyCore.ReciprocalLattice — Type

ReciprocalLattice(data::AbstractMatrix)

Construct a ReciprocalLattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Warning

Avoid using this constructor directly. Use reciprocal instead.

CrystallographyCore.basisvectors — Method

basisvectors(lattice::ReciprocalLattice)

Get the three basis vectors from a reciprocal lattice.

CrystallographyCore.reciprocal — Function

reciprocal(lattice::Lattice)
reciprocal(lattice::ReciprocalLattice)

Get the reciprocal of a Lattice or a ReciprocalLattice.

Cells

CrystallographyCore.Cell — Type

Cell(lattice, positions, atoms)

Create a new cell.

Argument lattice is a Lattice type. Fractional atomic positions positions are given by a vector of $N$ vectors with floating point values, where $N$ is the number of atoms. Argument atoms is a list of $N$ values, where the same kind of atoms need to be the same type.

CrystallographyCore.eachatom — Function

eachatom(cell::Cell)

Create a generator that iterates over the atoms in a Cell.